Set NPLC so samples are averaged over 50
This is a noise reduction measure, requested by the power supplies team

Also remove CLEAR; it isn't really necessary as we have to read all data
from the buffer anyway. It also seemed to cause an error on one device
I tested it against, as though it didn't understand the command.